requirements.txt
uvicorn[standard]==0.27.1
fastapi==0.109.2
python-multipart==0.0.9
pydantic==2.8.2
atlassian-python-api==3.41.19
alembic==1.13.1
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
pydantic-settings==2.3.4
prometheus_fastapi_instrumentator==6.1.0
jinja2==3.1.4
